+++ Test/UnitCommands/DeleteAllUnitTestDatabases.cs
@@ -1,6 +1,7 @@
 // Copyright (c) 2020 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT license. See License.txt in the project root for license information.
 
+using Test.Helpers;
 using TestSupport.Attributes;
 using TestSupport.EfHelpers;
 using Xunit.Abstractions;
@@ -16,22 +17,26 @@ namespace Test.UnitCommands
             _output = output;
         }
 
-        //Run this method to wipe ALL the test databases using your appsetting.json connection string
+        //Run this method to wipe ALL the SQL Server test databases using your appsetting.json connection string
         //You need to run it in debug mode - that stops it being run when you "run all" unit tests
         [RunnableInDebugOnly]  //#A
-        public void DeleteAllTestDatabasesOk() //#B
+        public void DeleteAllSqlServerTestDatabasesOk() //#B
         {
             var numDeleted = DatabaseTidyHelper //#C
                 .DeleteAllUnitTestDatabases();//#C
             _output.WriteLine(                              //#D
-                "This deleted {0} databases.", numDeleted); //#D
+                "This deleted {0} SQL Server databases.", numDeleted); //#D
         }
 
-        /****************************************************************
-        #A The [RunnableInDebugOnly] makes sure the unit command is not run by accident when the main unit tests are run. I must manually run this method in debug mode
-        #B This has the format of a unit test, that is a public method which returns void
-        #C I call the DeleteAllUnitTestBranchDatabases method from my EcCore.TestSupport library. This returns the number of databases that it deleted
-        #D I then write out how many database were deleted by this method
-         * ****************************************************************/
+        //Run this method to wipe ALL the PostgreSql test databases using your appsetting.json connection string
+        //You need to run it in debug mode - that stops it being run when you "run all" unit tests
+        [RunnableInDebugOnly]  //#A
+        public void DeleteAllPostgreSqlTestDatabasesOk()
+        {
+            var numDeleted = PostgreSqlHelpers
+                .DeleteAllPostgreSqlUnitTestDatabases();
+            _output.WriteLine( 
+                "This deleted {0} PostgreSql databases.", numDeleted);
+        }
     }
 }